SPLINES-01 - Stable version. - Modularization of methods.
